services/talk_booking/pyproject.toml
[tool.poetry]
name = "talk_booking"
version = "0.1.0"
description = ""
authors = ["Chase Frankenfeld <chasefrankenfeld@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
pydantic = {extras = ["email"], version = "^1.9.0"}
SQLAlchemy = "^1.4.35"
alembic = "^1.7.7"
boto3 = "^1.21.42"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
bandit = "^1.7.4"
safety = "^1.10.3"
requests = "^2.27.1"
pytest-env = "^0.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
